move tests for 'get_imdbid_from_uniqueids'
